Made minimal changes to make the crate to be both lib and bin
