+++ Cargo.toml
@@ -33,6 +33,14 @@ categories  = ["command-line-utilities", "encoding"]
 
 license     = "MIT"
 
+[[bin]]
+name = "rsbx"
+path = "src/bin/main.rs"
+
+[lib]
+name = "rsbx_lib"
+path = "src/lib.rs"
+
 [dependencies]
 reed-solomon-erasure = { version = "3.0", features = ["pure-rust"] }
 ring                 = "0.12.1"
